refactor: move ResultPage into result route
