+++ app/components/InfoBoxItem.tsx
@@ -5,10 +5,11 @@ import Image, { ImagePropsSchema } from "./Image";
 import RichText from "./RichText";
 import ButtonContainer from "./ButtonContainer";
 
+// Omitting React components to make schema serializable
 export const InfoBoxItemPropsSchema = z.object({
   identifier: z.string().optional(),
-  label: HeadingPropsSchema.optional(),
-  headline: HeadingPropsSchema.optional(),
+  label: HeadingPropsSchema.omit({ children: true }).optional(),
+  headline: HeadingPropsSchema.omit({ children: true }).optional(),
   image: ImagePropsSchema.optional(),
   content: z.string().optional(),
   button: ButtonPropsSchema.optional(),
