+++ app/routes/vorabcheck/result.tsx
@@ -6,7 +6,6 @@ import type {
 } from "@remix-run/node";
 import { json, redirect } from "@remix-run/node";
 import { getSession } from "~/sessions";
-import ResultPage from "~/components/ResultPage";
 import { getStrapiResultPage, getStrapiVorabCheckCommon } from "~/services/cms";
 import { buildFlowController } from "~/services/flow/buildFlowController";
 import { getReasonsToDisplay } from "~/models/flows/common";
@@ -15,7 +14,18 @@ import {
   flowSpecifics,
   splatFromParams,
 } from "./flowSpecifics";
-import type { InfoBoxItemProps } from "~/components/InfoBoxItem";
+import type { ReactElement } from "react";
+import CheckCircleOutline from "@mui/icons-material/CheckCircleOutline";
+import HighlightOff from "@mui/icons-material/HighlightOff";
+import WarningAmber from "@mui/icons-material/WarningAmber";
+import type { StrapiResultPageType } from "~/services/cms/models/StrapiResultPageType";
+import Container from "~/components/Container";
+import Heading from "~/components/Heading";
+import PageContent, { keyFromElement } from "~/components/PageContent";
+import RichText from "~/components/RichText";
+import InfoBox from "~/components/InfoBox";
+import { ProgressBar } from "~/components/form/ProgressBar";
+import { ButtonNavigation } from "~/components/form/ButtonNavigation";
 
 export const meta: V2_MetaFunction<typeof loader> = ({ data, location }) => [
   { title: data?.meta?.title ?? location.pathname },
@@ -34,14 +44,14 @@ export const loader = async ({ params, request }: LoaderArgs) => {
 
   // Slug change to keep Strapi slugs without ergebnis/
   const slug = pathname.replace(/ergebnis\//, "");
-  const resultContent = await getStrapiResultPage({ slug });
+  const content = await getStrapiResultPage({ slug });
   const reasonElementsWithID =
-    resultContent.reasonings.data?.map((el) => el.attributes) ?? [];
+    content.reasonings.data?.map((el) => el.attributes) ?? [];
 
   return json({
-    commonContent: await getStrapiVorabCheckCommon(),
-    resultContent,
-    meta: resultContent.meta,
+    common: await getStrapiVorabCheckCommon(),
+    content,
+    meta: content.meta,
     reasonsToDisplay: getReasonsToDisplay(reasonElementsWithID, data),
     progress: flowController.getProgress(stepId),
     previousStep: flowController.getPrevious(stepId)?.url,
@@ -57,23 +67,125 @@ export const action: ActionFunction = async ({ params, request }) => {
   return redirect(flowController.getNext("ergebnis/" + splat).url);
 };
 
+const iconCSS = "inline-block mr-8 !h-[36px] !w-[36px]";
+const icons: Record<StrapiResultPageType, ReactElement> = {
+  error: <HighlightOff color="error" className={`${iconCSS} !text-red-900`} />,
+  success: (
+    <CheckCircleOutline
+      color="success"
+      className={`${iconCSS} !text-green-900`}
+    />
+  ),
+  warning: (
+    <WarningAmber color="warning" className={`${iconCSS} !text-yellow-900`} />
+  ),
+};
+
+const backgrounds: Record<StrapiResultPageType, string> = {
+  error: "bg-red-200",
+  success: "bg-green-200",
+  warning: "bg-yellow-200",
+};
+
 export function Step() {
-  const {
-    commonContent,
-    resultContent,
-    reasonsToDisplay,
-    progress,
-    previousStep,
-  } = useLoaderData<typeof loader>();
+  const { common, content, reasonsToDisplay, progress, previousStep } =
+    useLoaderData<typeof loader>();
+
+  const documentsList = content.documents.data?.attributes.element ?? [];
+  const nextSteps = content.nextSteps.data?.attributes.element ?? [];
 
   return (
-    <ResultPage
-      content={resultContent}
-      common={commonContent}
-      backDestination={previousStep}
-      reasonsToDisplay={reasonsToDisplay as InfoBoxItemProps[]}
-      progressStep={progress.current}
-      progressTotal={progress.total}
-    />
+    <>
+      <div className={backgrounds[content.pageType]}>
+        <Container paddingTop="24">
+          <ProgressBar
+            label={common.progressBarLabel}
+            progress={progress.current}
+            max={progress.total}
+          />
+          <Heading
+            tagName={content.heading.tagName}
+            look={content.heading.look}
+            className="flex items-center mb-0"
+          >
+            {icons[content.pageType]}
+            {content.heading.text}
+          </Heading>
+        </Container>
+
+        {content.hintText && (
+          <Container
+            backgroundColor="white"
+            paddingTop="32"
+            paddingBottom="40"
+            overhangingBackground={true}
+          >
+            <div className="ds-stack-8">
+              <p className="ds-label-02-bold">{common.resultHintLabel}</p>
+              <RichText markdown={content.hintText.text} />
+            </div>
+          </Container>
+        )}
+
+        {content.linkText && (
+          <Container paddingTop="32" paddingBottom="32">
+            <a
+              href={"/beratungshilfe/vorabcheck"}
+              className="block ds-link-02-bold"
+            >
+              {content.linkText}
+            </a>
+          </Container>
+        )}
+      </div>
+      {content.freeZone.length > 0 && (
+        <Container>
+          <PageContent content={content.freeZone} />
+        </Container>
+      )}
+      {reasonsToDisplay.length > 0 && (
+        <Container>
+          <InfoBox
+            heading={{
+              tagName: "h2",
+              look: "ds-heading-02-reg",
+              text: "Begründung",
+              className: "mb-16",
+            }}
+            items={reasonsToDisplay}
+          />
+        </Container>
+      )}
+
+      {documentsList.length > 0 && (
+        <div>
+          {documentsList.map((element, idx) => (
+            <div key={keyFromElement(element)}>
+              <PageContent content={[element]} />
+              {idx != 0 && idx != documentsList.length - 1 && (
+                <hr className="my-24" />
+              )}
+            </div>
+          ))}
+        </div>
+      )}
+      <div className={`${documentsList.length > 0 && "bg-blue-100"}`}>
+        <Container>
+          <ButtonNavigation
+            back={{
+              destination: previousStep,
+              label: common.backButtonDefaultLabel,
+            }}
+            next={{
+              destination: content.nextLink?.url ?? undefined,
+              label: content.nextLink?.text ?? undefined,
+            }}
+          />
+        </Container>
+        <div className="pb-48">
+          <PageContent content={nextSteps} />
+        </div>
+      </div>
+    </>
   );
 }
